mouse-sensor-toys/generate-anoto/pyproject.toml
[project]
name = "generate-anoto"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "microdots",
    "numpy>=2.2.3",
    "pillow>=11.1.0",
]

[tool.uv.sources]
microdots = { git = "https://github.com/cheind/py-microdots" }
